+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ------------------------------------------------------------------------
 - Feature: [#2697] Add initial OpenGraphics custom assets.
 - Fix: [#2676] Cargo labels are misaligned in vehicle window.
+- Fix: [#2690] Inability to remove certain track pieces.
 - Fix: [#2691] Scenarios with original procedural terrain generation could crash.
+- Fix: [#2692] Inability to place certain track pieces.
 
 24.10 (2024-10-20)
 ------------------------------------------------------------------------